requirements.txt
uvicorn
fastapi
pymysql
sqlalchemy
alembic
databases
fastapi_pagination
passlib
python-jose
python-multipart